+++ src/dns_encode.erl
@@ -596,8 +596,8 @@ encode_rrdata(
     {<<Priority:16, Weight:16, Target/binary>>, CompMap};
 encode_rrdata(_Pos, _Class, #dns_rrdata_resinfo{data = Strings}, CompMap) ->
     {encode_text(Strings), CompMap};
-encode_rrdata(_Pos, ?DNS_CLASS_IN, #dns_rrdata_wallet{data = Bin}, CompMap) ->
-    {Bin, CompMap};
+encode_rrdata(_Pos, ?DNS_CLASS_IN, #dns_rrdata_wallet{data = Strings}, CompMap) ->
+    {encode_text(Strings), CompMap};
 encode_rrdata(
     _Pos,
     _Class,